requirements.txt
streamlit==1.49.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.32.3